packages/api/pyproject.toml
[project]
name = "taskflow-api"
version = "0.1.0"
description = "TaskFlow Backend API - Human-Agent Task Management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/taskflow_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/taskflow_api/tests"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
